List heats for a given event
The rounds typically undertaken that we're interested in operate in
a series of 'heats', the results of which are combined to create a
final result for the event. Now we can get all the heats for a
given event, and pump those into get_ordered_heat_results.

Also fixes an issue where some competitors came back in the results
list but did not actually compete.