Support tie-breaks in round standings
As a round is made up of multiple heats, it is very possible for
two competitors to have the same points when their heat results
are pulled together and assigned points. Thus, we need to have a
tie-break mechanism.

A tie is broken based on the competitors' best
individual results. E.g. a competitor placing first, third and
fifth would be ahead of a competitor placing second, second and
fourth.

Tie breaks cascade, such that a competitor placing first, third and
fifth would be ahead of a competitor placing first, fourth and
fourth.

Currently unsupported are cases when all results are the same. As
it stands, Python will handle that, which I suspect means sorting
by competitor name alphabetically, but I'd need to test it.

The code is a bit messy as it stands, but it works, so acts as a
solid baseline to build any refactors from.